pyproject.toml
[project]
name = "towercellsrd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==6.0.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "drf-spectacular-sidecar>=2025.12.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "celery>=5.4.0",
    "redis>=5.2.1",
    "psycopg[binary]>=3.3.2",
]
